Update internal classes
